refactor mcp path expr parser
